requirements.txt
asgiref==3.8.1
cachetools==5.5.0
distlib==0.3.8
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
Markdown==3.7
platformdirs==4.3.6
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.5
